requirements/prod.txt
aiogram==3.21.0
openai==1.98.0
python-dotenv==1.1.1
